Add a schema file for the configuration file (#435)
* feat: Create initial schema file to use for validating configuration files.

Updated the dataclasses with annotations to provide descriptions.

* refactor: Update the schema generation to list enums as valid options for the serial config options.

Removed the requirement for the type and title on every property, since that data just got in the way when using the IDE auto-complete feature.

* test: Added tests to the schema generation script.

Also updated docs to indicate that the address of a device should be expressed as a string in the config file.

* ci: Added a pre-commit hook to generate the schema every time a python file is modified

* test: Added a test file to verify the schema correctly validates config files.

Also, added a step to check that the generated schema has valid links

* chore: Update dependencies in pyproject.toml to include dc-schema and jsonschema

* chore: Cleaned up all failing linting checks

* ci: Skip a local pre-commit hook when running via pre-commit.ci